pyproject.toml
[tool.poetry]
name = "nonebot-adapter-onebot-pretender"
version = "0.1.0"
description = ""
authors = ["ssttkkl <huang.wen.long@hotmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/ssttkkl/nonebot-adapter-onebot-pretender"
repository = "https://github.com/ssttkkl/nonebot-adapter-onebot-pretender"
documentation = ""
keywords = []
classifiers = []
packages = [
  { include = "nonebot_adapter_onebot_pretender", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.1.0"
nonebot-adapter-onebot = "^2.3.0"
nonebot-plugin-datastore = "^1.1.2"
unqlite = "^0.9.6"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.275"
isort = "^5.10.1"
black = "^23.1.0"
nonemoji = "^0.1.2"
pre-commit = "^3.1.0"
nonebot2 = { extras = ["fastapi", "websockets", "httpx"], version = "^2.1.0" }


[tool.poetry.group.adapters.dependencies]
nonebot-adapter-red = "^0.5.0"


[tool.poetry.group.test_plugins.dependencies]
nonebot-plugin-escape-url = { version = "*", python = "^3.9" }
nonebot-plugin-access-control = { version = "*", python = "^3.9" }
nonebot-plugin-pixivbot = { version = ">= 2.0.4", python = "^3.9" }
nonebot-plugin-mahjong-utils = { version = ">= 0.4.3", extras = ["htmlrender"], python = "^3.9" }

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 88
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[tool.ruff]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["E402", "C901", "UP037", "PYI021"]

line-length = 88
target-version = "py38"

[tool.ruff.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.nonebot]
plugins = [
  "nonebot_plugin_escape_url",
  "nonebot_plugin_access_control",
  "nonebot_plugin_mahjong_utils",
  "nonebot_plugin_pixivbot"
]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
